requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
line-bot-sdk==3.3.0
python-dotenv==1.0.0
supabase==2.0.3
python-dateutil==2.8.2 